backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.7
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
whisper==1.1.10
torch>=2.0.0
torchaudio>=2.0.0
soundfile==0.12.1
numpy>=1.24.0
chromadb==0.4.18
sentence-transformers==2.2.2
